requirements.txt
streamlit
folium
streamlit-folium
google-generativeai
pandas
geopy
deep-translator